HomogeneousData

ExtendableFEM.HomogeneousData — Method

function HomogeneousData(u; entities = ON_CELLS, kwargs...)

When assembled, the unknown u of the Problem will be penalized to zero on the specifies entities and entity regions (via kwargs).

Keyword arguments:

regions: subset of regions where operator should be assembly only. Default: Any[]

verbosity: verbosity level. Default: 0

name: name for operator used in printouts. Default: ''HomogeneousData''

penalty: penalty for fixed degrees of freedom. Default: 1.0e30

mask: array of zeros/ones to set which components should be set by the operator (only works with componentwise dofs, add a 1 or 0 to mask additional dofs). Default: Any[]

ExtendableFEM.HomogeneousBoundaryData — Method

function HomogeneousBoundaryData(u; entities = ON_BFACES, kwargs...)

When assembled, the unknown u of the Problem will be penalized to zero on the boundary faces and boundary regions (via kwargs).

Keyword arguments:

regions: subset of regions where operator should be assembly only. Default: Any[]

verbosity: verbosity level. Default: 0

name: name for operator used in printouts. Default: ''HomogeneousData''

penalty: penalty for fixed degrees of freedom. Default: 1.0e30

mask: array of zeros/ones to set which components should be set by the operator (only works with componentwise dofs, add a 1 or 0 to mask additional dofs). Default: Any[]