InterpolateBoundaryData

ExtendableFEM.InterpolateBoundaryData — Type

function InterpolateBoundaryData(u, data!::Function; kwargs...)

When assembled, the unknown u of the Problem will be penalized to match the standard interpolation of the provided data! function. The header of this function needs to be conform to the interface

data!(result, qpinfo)

where qpinfo allows to access information at the current quadrature point, e.g. qpinfo.x provides the global coordinates of the quadrature/evaluation point.

Keyword arguments:

params: array of parameters that should be made available in qpinfo argument of kernel function. Default: nothing

regions: subset of regions where operator should be assembly only. Default: Any[]

name: name for operator used in printouts. Default: ''BoundaryData''

bonus_quadorder: additional quadrature order added to the quadorder chosen by the interpolator. Default: 0

penalty: penalty for fixed degrees of freedom. Default: 1.0e30

verbosity: verbosity level. Default: 0